jna.nosys should not be forced to "true"
jna.nosys can be enabled by the host program or by commandline parameters. With the current implementation it cannot be disabled.
This forces to enable DLL unpacking and loading from %TEMP%, instead of providing a location from the host, e.g. via java.library.path. Loading a library from %TEMP% is a security issue and should not be done by modern applications.